Fixes test suite to run with newer Minitest
Fixes order cancellation API calls

Parameters being sent to the orders API were being
attached to the query string instead of added to the body.

As a result their truthiness was being lost and all calls to
cancel an order would result in an email being delivered to
the customer.

Ensures consistent state for tests

There was an intermittent bug that would come up in some
tests because the ShopifyAPI::Session.secret wasn't set.